+++ riskgpt/utils/__init__.py
@@ -1,0 +1,19 @@
+"""Utility modules for RiskGPT."""
+
+from riskgpt.utils.circuit_breaker import (
+    duckduckgo_breaker,
+    openai_breaker,
+    with_fallback,
+)
+from riskgpt.utils.memory_factory import get_memory, register_memory_backend
+from riskgpt.utils.prompt_loader import load_prompt, load_system_prompt
+
+__all__ = [
+    "duckduckgo_breaker",
+    "get_memory",
+    "load_prompt",
+    "load_system_prompt",
+    "openai_breaker",
+    "register_memory_backend",
+    "with_fallback",
+]